Move rules from registration factory to base plugin
