+++ subprojects/platform-jni/src/main/java/dev/nokee/platform/jni/internal/ApiElementsConfiguration.java
@@ -21,7 +21,7 @@ import dev.nokee.model.internal.core.ModelNode;
 public final class ApiElementsConfiguration implements ModelComponent {
 	private final ModelNode entity;
 
-	ApiElementsConfiguration(ModelNode entity) {
+	public ApiElementsConfiguration(ModelNode entity) {
 		this.entity = entity;
 	}
